+++ test/ft_ark_a2preservation_test.rb
@@ -1,41 +1,33 @@
 # frozen_string_literal: true
 require 'webmock/minitest'
 require_relative './test_helper'
-require_relative '../lib/fair_tests/ft_ark_f1gupr'
+require_relative '../lib/fair_tests/ft_ark_a2preservation'
 
-class FtArkF1Test < Minitest::Test
+class FtArkA2preservationTest < Minitest::Test
   include ::TestHelper
-  include ::FtArkF1
+  include ::FtArkA2preservation
 
-  def test_pass_ft_ark_f1gupr1
+  def test_pass_ft_ark_a2preservation
     stub_request(:post, "#{ENV['FAIRSHARING_API_URL']}").
       with(headers: headers).to_return(
       status: 200,
       body: {
         "data": {
           "fairsharingRecord": {
-            "id": "1234567",
+            "id": "123456",
             "registry": "Database",
-            "recordAssociations": [
-              {
-                "recordAssocLabel": "implements",
-                "linkedRecord": {
-                  "type": "identifier_schema",
-                  "metadata": {
-                    "persistent": true,
-                    "globally_unique": true,
-                    "resolvable": true
-                  }
-                }
+            "metadata": {
+              "data_preservation_policy": {
+                "url": "https://www.what_an_url.klo"
               }
-            ]
+            }
           }
         }
       }.to_json,
       headers: headers
     )
 
-    post '/test/ft_ark_f1gupr1',
+    post '/test/ft_ark_a2preservation',
          params: { resource_identifier: 'https://fairsharing.org/1234' }.to_json,
          headers: headers
 
@@ -45,34 +37,27 @@ class FtArkF1Test < Minitest::Test
     assert_equal body['value'], 'pass'
   end
 
-  def test_fail_ft_ark_f1gupr1
+  def test_fail_ft_ark_a2preservation
     stub_request(:post, "#{ENV['FAIRSHARING_API_URL']}").
       with(headers: headers).to_return(
       status: 200,
       body: {
         "data": {
           "fairsharingRecord": {
-            "id": "1234567",
+            "id": "123456",
             "registry": "Database",
-            "recordAssociations": [
-              {
-                "recordAssocLabel": "implements",
-                "linkedRecord": {
-                  "type": "identifier_schema",
-                  "metadata": {
-                    "persistent": true,
-                    "globally_unique": false
-                  }
-                }
+            "metadata": {
+              "data_preservation_policy": {
+                "url": ""
               }
-            ]
+            }
           }
         }
       }.to_json,
       headers: headers
     )
 
-    post '/test/ft_ark_f1gupr1',
+    post '/test/ft_ark_a2preservation',
          params: { resource_identifier: 'https://fairsharing.org/1234' }.to_json,
          headers: headers
 
@@ -82,7 +67,7 @@ class FtArkF1Test < Minitest::Test
     assert_equal body['value'], 'fail'
   end
 
-  def test_indeterminate_ft_ark_f1gupr1
+  def test_indeterminate_ft_ark_a2preservation
     stub_request(:post, "#{ENV['FAIRSHARING_API_URL']}").
       with(headers: headers).to_return(
       status: 200,
@@ -96,7 +81,7 @@ class FtArkF1Test < Minitest::Test
       headers: headers
     )
 
-    post '/test/ft_ark_f1gupr1',
+    post '/test/ft_ark_a2preservation',
          params: { resource_identifier: 'https://fairsharing.org/1234' }.to_json,
          headers: headers
