+++ test/ft_ark_f1gupr_test.rb
@@ -1,33 +1,41 @@
 # frozen_string_literal: true
 require 'webmock/minitest'
 require_relative './test_helper'
-require_relative '../lib/fair_tests/ft_ark_a2preservation'
+require_relative '../lib/fair_tests/ft_ark_f1gupr'
 
-class FtArkF1Test < Minitest::Test
+class FtArkf1guprTest < Minitest::Test
   include ::TestHelper
-  include ::FtArkF1
+  include ::FtArkF1gupr
 
-  def test_pass_ft_ark_a2preservation
+  def test_pass_ft_ark_f1gupr
     stub_request(:post, "#{ENV['FAIRSHARING_API_URL']}").
       with(headers: headers).to_return(
       status: 200,
       body: {
         "data": {
           "fairsharingRecord": {
-            "id": "123456",
+            "id": "1234567",
             "registry": "Database",
-            "metadata": {
-              "data_preservation_policy": {
-                "url": "https://www.what_an_url.klo"
+            "recordAssociations": [
+              {
+                "recordAssocLabel": "implements",
+                "linkedRecord": {
+                  "type": "identifier_schema",
+                  "metadata": {
+                    "persistent": true,
+                    "globally_unique": true,
+                    "resolvable": true
+                  }
+                }
               }
-            }
+            ]
           }
         }
       }.to_json,
       headers: headers
     )
 
-    post '/test/ft_ark_a2preservation',
+    post '/test/ft_ark_f1gupr',
          params: { resource_identifier: 'https://fairsharing.org/1234' }.to_json,
          headers: headers
 
@@ -37,27 +45,34 @@ class FtArkF1Test < Minitest::Test
     assert_equal body['value'], 'pass'
   end
 
-  def test_fail_ft_ark_a2preservation
+  def test_fail_ft_ark_f1gupr
     stub_request(:post, "#{ENV['FAIRSHARING_API_URL']}").
       with(headers: headers).to_return(
       status: 200,
       body: {
         "data": {
           "fairsharingRecord": {
-            "id": "123456",
+            "id": "1234567",
             "registry": "Database",
-            "metadata": {
-              "data_preservation_policy": {
-                "url": ""
+            "recordAssociations": [
+              {
+                "recordAssocLabel": "implements",
+                "linkedRecord": {
+                  "type": "identifier_schema",
+                  "metadata": {
+                    "persistent": true,
+                    "globally_unique": false
+                  }
+                }
               }
-            }
+            ]
           }
         }
       }.to_json,
       headers: headers
     )
 
-    post '/test/ft_ark_a2preservation',
+    post '/test/ft_ark_f1gupr',
          params: { resource_identifier: 'https://fairsharing.org/1234' }.to_json,
          headers: headers
 
@@ -67,7 +82,7 @@ class FtArkF1Test < Minitest::Test
     assert_equal body['value'], 'fail'
   end
 
-  def test_indeterminate_ft_ark_a2preservation
+  def test_indeterminate_ft_ark_f1gupr
     stub_request(:post, "#{ENV['FAIRSHARING_API_URL']}").
       with(headers: headers).to_return(
       status: 200,
@@ -81,7 +96,7 @@ class FtArkF1Test < Minitest::Test
       headers: headers
     )
 
-    post '/test/ft_ark_a2preservation',
+    post '/test/ft_ark_f1gupr',
          params: { resource_identifier: 'https://fairsharing.org/1234' }.to_json,
          headers: headers
